requirements.txt
cloudshell-shell-core
cloudshell-orch-core